+++ packages/slots/test/examples.browser.test.tsx
@@ -38,7 +38,7 @@ describe('@bento/slots examples', function bento() {
       };
 
       assume(result).equals(
-        '<div data-override="slot" id=":r0:" class="example"><label for="example">Hello World</label><button id=":r1:" data-override="slot">Click Me</button></div>'
+        '<div data-override="slot" data-slot="example-container" id=":r0:" class="example"><label for="example" data-slot="label">Hello World</label><button id=":r1:" data-override="slot" data-slot="button">Click Me</button></div>'
       );
       await screen.getByRole('button', { name: 'Click me' }).click();
 
@@ -56,7 +56,7 @@ describe('@bento/slots examples', function bento() {
       const result = container.innerHTML;
 
       assume(result).equals(
-        '<div class="example"><label for="example"><strong>Hello World</strong></label><button id=":r2:">Click Me</button></div>'
+        '<div data-slot="example-container" class="example"><label for="example" data-slot="label"><strong>Hello World</strong></label><button id=":r2:" data-slot="button">Click Me</button></div>'
       );
     });
   });
@@ -73,7 +73,7 @@ describe('@bento/slots examples', function bento() {
       };
 
       assume(result).equals(
-        '<div class="example"><label for="example">Hello World</label><button id=":r3:" data-override="style slot" style="background: red; border: 2px solid black;">Click Me</button></div>'
+        '<div data-slot="example-container" class="example"><label for="example" data-slot="label">Hello World</label><button id=":r3:" data-override="style slot" data-slot="button" style="background: red; border: 2px solid black;">Click Me</button></div>'
       );
       await screen.getByRole('button', { name: 'Click me' }).click();
 
@@ -91,9 +91,9 @@ describe('@bento/slots examples', function bento() {
 
       assume(result).equals(
         [
-          '<p>',
+          '<p data-slot="description">',
           '  <span>Description: </span>',
-          '  <label data-override="className slot" class="merged-class" title="better-title" id="merged">',
+          '  <label data-override="className slot" data-slot="label" class="merged-class" title="better-title" id="merged">',
           '  </label>',
           '  <span>Merged summary</span>',
           '  <span>Better error</span>',
@@ -128,8 +128,8 @@ describe('@bento/slots examples', function bento() {
       const result = container.innerHTML;
 
       assume(result).equals(
-        '<label data-override="className slot" class="title-1">title</label>' +
-          '<label data-override="className slot" class="description-1">title</label>'
+        '<label data-override="className slot" data-slot="title" class="title-1">title</label>' +
+          '<label data-override="className slot" data-slot="title" class="description-1">title</label>'
       );
     });
 
@@ -138,8 +138,8 @@ describe('@bento/slots examples', function bento() {
       const result = container.innerHTML;
 
       assume(result).equals(
-        '<label data-override="className slot" class="title-2">title</label>' +
-          '<label data-override="className slot" class="description-2">title</label>'
+        '<label data-override="className slot" data-slot="title" class="title-2">title</label>' +
+          '<label data-override="className slot" data-slot="title" class="description-2">title</label>'
       );
     });
 
@@ -148,8 +148,8 @@ describe('@bento/slots examples', function bento() {
       const result = container.innerHTML;
 
       assume(result).equals(
-        '<label data-override="className slot" class="title-3">title</label>' +
-          '<label data-override="className slot" class="description-3">title</label>'
+        '<label data-override="className slot" data-slot="title" class="title-3">title</label>' +
+          '<label data-override="className slot" data-slot="title" class="description-3">title</label>'
       );
     });
 
@@ -160,7 +160,8 @@ describe('@bento/slots examples', function bento() {
       console.log('bbbb', result);
 
       assume(result).equals(
-        '<div>' + '<label data-override="className slot" class="inherited-title">title</label>Content' + '</div>'
+        '<div>' +
+          '<label data-override="className slot" data-slot="title" class="inherited-title">title</label>Content</div>'
       );
     });
   });